+++ src/kaocha/api.clj
@@ -3,6 +3,7 @@
   "Programmable test runner interface."
   (:require [clojure.test :as t]
             [kaocha.config :as config]
+            [kaocha.hierarchy :as hierarchy]
             [kaocha.history :as history]
             [kaocha.output :as output]
             [kaocha.plugin :as plugin]
@@ -97,11 +98,10 @@
           (with-bindings (config/binding-map config)
             (let [config (resolve-reporter config)]
               (let [test-plan (test-plan config)]
-                (when (empty? (:kaocha.test-plan/tests test-plan))
+                (when-not (some hierarchy/leaf? (testable/test-seq test-plan))
                   (output/warn (str "No tests were found, make sure :test-paths and "
                                     ":ns-patterns are configured correctly in tests.edn."))
                   (throw+ {:kaocha/early-exit 0}))
-
                 (when (find-ns 'matcher-combinators.core)
                   (require 'kaocha.matcher-combinators))